+++ CHANGELOG.md
@@ -11,10 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - `list.h`:
    - `ML99_listFromTuples` to transform comma-separated tuples into a list.
- - `util.h`:
+ - New module `ident.h`:
+   - Migrate `ML99_detectIdent`, `ML99_identEq`, `ML99_DETECT_IDENT`, `ML99_IDENT_EQ`, `ML99_C_KEYWORD_DETECTOR`, `ML99_UNDERSCORE_DETECTOR` from `util.h`.
    - `ML99_LOWERCASE_DETECTOR` to detect lowercase characters.
    - `ML99_UPPERCASE_DETECTOR` to detect uppercase characters.
    - `ML99_DIGIT_DETECTOR` to detect digits.
+
+### Changed
+
+ - `util.h`:
+   - Automatically include `ident.h` for backwards compatibility.
 
 ### Fixed